# services/vehicle-service/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vehicle Inventory Management API",
    "description": "Comprehensive 2-Wheeler EV Fleet Management System",
    "version": "1.0.0",
    "contact": {
      "name": "EV91 Platform",
      "email": "api@ev91.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4003",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/vehicles": {
      "get": {
        "summary": "Get all vehicles with filtering and pagination",
        "tags": ["Vehicles"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"in": "query", "name": "page", "schema": {"type": "integer", "default": 1}},
          {"in": "query", "name": "limit", "schema": {"type": "integer", "default": 20}},
          {"in": "query", "name": "oemType", "schema": {"type": "string"}},
          {"in": "query", "name": "operationalStatus", "schema": {"type": "string"}},
          {"in": "query", "name": "assignedRider", "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "List of vehicles"}
        }
      },
      "post": {
        "summary": "Create a new vehicle",
        "tags": ["Vehicles"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CreateVehicleRequest"}
            }
          }
        },
        "responses": {
          "201": {"description": "Vehicle created successfully"},
          "400": {"description": "Validation error"},
          "401": {"description": "Unauthorized"}
        }
      }
    },
    "/api/v1/vehicles/stats": {
      "get": {
        "summary": "Get vehicle statistics",
        "tags": ["Vehicles"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Vehicle statistics including counts by status, model distribution, etc."}
        }
      }
    },
    "/api/v1/vehicles/{id}": {
      "get": {
        "summary": "Get vehicle by ID",
        "tags": ["Vehicles"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"in": "path", "name": "id", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Vehicle details"},
          "404": {"description": "Vehicle not found"}
        }
      },
      "put": {
        "summary": "Update vehicle",
        "tags": ["Vehicles"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"in": "path", "name": "id", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/UpdateVehicleRequest"}
            }
          }
        },
        "responses": {
          "200": {"description": "Vehicle updated successfully"},
          "404": {"description": "Vehicle not found"}
        }
      },
      "delete": {
        "summary": "Delete vehicle (soft delete)",
        "tags": ["Vehicles"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"in": "path", "name": "id", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Vehicle deleted successfully"},
          "404": {"description": "Vehicle not found"}
        }
      }
    },
    "/api/v1/vehicles/{id}/status": {
      "get": {
        "summary": "Get vehicle status",
        "tags": ["Vehicles"],
        "parameters": [
          {"in": "path", "name": "id", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Vehicle status"}
        }
      },
      "patch": {
        "summary": "Update vehicle status",
        "tags": ["Vehicles"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"in": "path", "name": "id", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "operationalStatus": {"type": "string", "enum": ["Available", "Assigned", "Under Maintenance", "Retired", "Damaged"]},
                  "serviceStatus": {"type": "string", "enum": ["Active", "Inactive", "Scheduled for Service"]},
                  "reason": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Status updated successfully"}
        }
      }
    },
    "/api/v1/vehicles/{id}/assign": {
      "post": {
        "summary": "Assign vehicle to rider",
        "tags": ["Vehicles"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"in": "path", "name": "id", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "riderId": {"type": "string"},
                  "assignmentNotes": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Vehicle assigned successfully"}
        }
      }
    },
    "/api/v1/vehicles/{id}/unassign": {
      "post": {
        "summary": "Unassign vehicle from rider",
        "tags": ["Vehicles"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"in": "path", "name": "id", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Vehicle unassigned successfully"}
        }
      }
    },
    "/api/v1/vehicles/{id}/history": {
      "get": {
        "summary": "Get vehicle history (status changes, assignments, etc.)",
        "tags": ["Vehicles"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"in": "path", "name": "id", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Vehicle history"}
        }
      }
    }
  },
...
}